+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added some help to discover the error cause when `New-LabAzureResourceGroupDeployment` failed
 - Added stored procedures for DSC database cleanup
 - Improved the discovery of Azure role sizes
+- Added error handling to 'Connect-LWAzureLabSourcesDrive'
 
 ### Bugs